keymatrix: correct header for rows change

+++ keymatrix.h
@@ -1,6 +1,6 @@
 struct keymatrix;
 
-uint8_t keymatrix_input(struct keymatrix *km, uint8_t scanbits);
+uint8_t keymatrix_input(struct keymatrix *km, uint16_t scanbits);
 bool keymatrix_SDL2event(struct keymatrix *km, SDL_Event *ev);
 void keymatrix_free(struct keymatrix *km);
 struct keymatrix *keymatrix_create(unsigned int rows, unsigned int cols, SDL_Keycode *matrix);
